speech-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
openai-whisper==20231117
torch==2.1.1
numpy==1.24.3
python-multipart==0.0.6
